Flytter logging funksjoner til eget util fil.
