+++ package.json
@@ -30,6 +30,7 @@
     "fp-ts": "^2.16.1",
     "io-ts": "^2.2.20",
     "next": "13.4.19",
+    "next-logger": "^3.0.2",
     "pino": "^8.15.1",
     "pino-pretty": "^10.2.0",
     "postcss": "8.4.29",
